CORE-944: Compiler Warnings

+++ WalletKitCore/src/ripple/BRRippleAddress.c
@@ -143,7 +143,7 @@ BRRippleAddress rippleAddressStringToAddress(const char* input)
 
     // Decode the string input
     uint8_t bytes[25];
-    int length = BRBase58DecodeEx(NULL, 0, input, rippleAlphabet);
+    size_t length = BRBase58DecodeEx(NULL, 0, input, rippleAlphabet);
     if (length != 25) {
         // Since ripple addresses are created from 20 byte account IDs the
         // needed space to covert it back has to be 25 bytes.
